Add simpler proof for commuting probability.
Since we already proved this theorem by Frobenius, we have a very easy
way to show the formula for the commuting probability in terms of the
number of conjugacy classes.

It might also be interesting to make a remark about the "Joseph"
conjectures, the last of which was only solved in 2022!. They are
statements about the values cp(G) can take. Here is the link to the
paper solving the last of the 3 statements:

https://arxiv.org/abs/2201.09402